+++ app/components/primer/link_component.rb
@@ -18,7 +18,7 @@ module Primer
       @system_arguments[:href] = href
       @system_arguments[:classes] = class_names(
         @system_arguments[:classes],
-        "muted-link" => fetch_or_fallback([true, false], muted, false)
+        "muted-link" => fetch_or_fallback_boolean(muted, false)
       )
     end